requirements.txt
altair==5.5.0
anchor_exp==0.0.2.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lime==0.2.0.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
murmurhash==1.0.12
narwhals==1.34.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
preshed==3.0.9
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.44.1
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.3.30
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
wrapt==1.17.2
